chore: moved dsl to old package to integrate the new one
